+++ src/main/java/dev/kingtux/tms/mlayout/IKeyBinding.java
@@ -68,4 +68,6 @@ public interface IKeyBinding {
     void tms$addAlternative(KeyBinding binding);
 
     int tms$getIndexInParent();
+    /// Proper Debug Message for KeyBinding
+    String tms$debugString();
 }