+++ src/main/kotlin/dev/kingtux/tms/alternatives/AlternativeKeyBinding.kt
@@ -46,4 +46,8 @@ class AlternativeKeyBinding(
     override fun isDefault(): Boolean {
         return this.defaultKey == this.boundKey
     }
+
+    override fun toString(): String {
+        return "AlternativeKeyBinding{parent=${this.translationKey}, key=${this.boundKey}, modifiers=${(this as IKeyBinding).`tms$getKeyModifiers`()}}"
+    }
 }